Make jumping sound work properly, but you can still spam an infinite amount which is not what is wanted.
This update gets the jump sound to actually work, but you can spam an infinite amount which is not intended, and so I am asking for help on how to fix that.

I also added numbers to indicate how many coins and papers you have, as well as changed the healthbar face to wince in pain when you touch an enemy.

I finally also added basic enemy AI, that only moves back and forth without paying attention to their surroundings, but I am going to be working on realistic enemy AI after I have ironed out the infinite jump bug, as well as figured out how to reset timers properly so that the timers don't accumulate each time they are called like the wincing face one does currently.